+++ ElevateEngine/src/ElevateEngine/Serialization/Serializer.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <ElevateEngine/Serialization/ObjectPropertyField.h>
+#include <ElevateEngine/Serialization/PropertyField.h>
 #include <ElevateEngine/Core/Byte.h>
 
 namespace Elevate
